+++ carp_firebase_backend/lib/carp_firebase_backend.dart
@@ -8,6 +8,7 @@ import 'package:firebase_storage/firebase_storage.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:json_annotation/json_annotation.dart';
+import 'package:carp_core/carp_core.dart';
 
 part 'firebase/firebase_data_endpoint.dart';
 part 'firebase/firebase_data_manager.dart';