+++ carp_mobile_sensing/lib/carp_mobile_sensing.dart
@@ -5,7 +5,6 @@ library carp_mobile_sensing;
 // The core packages
 export 'package:carp_mobile_sensing/datastore/datastore.dart';
 export 'package:carp_mobile_sensing/runtime/runtime.dart';
-export 'package:carp_mobile_sensing/domain/domain.dart';
 
 // The list of built-in sensors
 export 'package:carp_mobile_sensing/probes/location/location.dart';