+++ carp_mobile_sensing/lib/probes/apps/apps.dart
@@ -5,6 +5,7 @@ import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:device_apps/device_apps.dart';
 import 'dart:async';
+import 'package:carp_core/carp_core.dart';
 
 part 'apps_datum.dart';
 part 'apps_probe.dart';