+++ carp_mobile_sensing/lib/probes/communication/communication.dart
@@ -8,6 +8,7 @@ import 'package:json_annotation/json_annotation.dart';
 import 'package:sms/sms.dart';
 import 'dart:async';
 import 'package:call_log/call_log.dart';
+import 'package:carp_core/carp_core.dart';
 
 part 'communication_datum.dart';
 part 'communication_measures.dart';