+++ carp_mobile_sensing/lib/probes/connectivity/connectivity.dart
@@ -8,6 +8,7 @@ import 'package:flutter_blue/flutter_blue.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:connectivity/connectivity.dart';
+import 'package:carp_core/carp_core.dart';
 
 part 'connectivity_probe.dart';
 part 'bluetooth_probe.dart';