+++ carp_mobile_sensing/lib/probes/sensors/sensors.dart
@@ -12,6 +12,7 @@ import 'package:light/light.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:pedometer/pedometer.dart';
 import 'package:stats/stats.dart';
+import 'package:carp_core/carp_core.dart';
 
 part 'sensor_probes.dart';
 part 'light_probe.dart';