+++ carp_mobile_sensing/lib/probes/sensors/sensors.g.dart
@@ -8,7 +8,7 @@ part of sensors;
 
 PedometerDatum _$PedometerDatumFromJson(Map<String, dynamic> json) {
   return PedometerDatum()
-    ..classname = json['classname'] as String
+    ..c__ = json['c__'] as String
     ..id = json['id'] as String
     ..timestamp = json['timestamp'] == null
         ? null
@@ -34,7 +34,7 @@ Map<String, dynamic> _$PedometerDatumToJson(PedometerDatum instance) {
     }
   }
 
-  writeNotNull('classname', instance.classname);
+  writeNotNull('c__', instance.c__);
   writeNotNull('id', instance.id);
   writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('device_info', instance.deviceInfo);
@@ -49,7 +49,7 @@ AccelerometerDatum _$AccelerometerDatumFromJson(Map<String, dynamic> json) {
       x: (json['x'] as num)?.toDouble(),
       y: (json['y'] as num)?.toDouble(),
       z: (json['z'] as num)?.toDouble())
-    ..classname = json['classname'] as String
+    ..c__ = json['c__'] as String
     ..id = json['id'] as String
     ..timestamp = json['timestamp'] == null
         ? null
@@ -68,7 +68,7 @@ Map<String, dynamic> _$AccelerometerDatumToJson(AccelerometerDatum instance) {
     }
   }
 
-  writeNotNull('classname', instance.classname);
+  writeNotNull('c__', instance.c__);
   writeNotNull('id', instance.id);
   writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('device_info', instance.deviceInfo);
@@ -83,7 +83,7 @@ GyroscopeDatum _$GyroscopeDatumFromJson(Map<String, dynamic> json) {
       x: (json['x'] as num)?.toDouble(),
       y: (json['y'] as num)?.toDouble(),
       z: (json['z'] as num)?.toDouble())
-    ..classname = json['classname'] as String
+    ..c__ = json['c__'] as String
     ..id = json['id'] as String
     ..timestamp = json['timestamp'] == null
         ? null
@@ -102,7 +102,7 @@ Map<String, dynamic> _$GyroscopeDatumToJson(GyroscopeDatum instance) {
     }
   }
 
-  writeNotNull('classname', instance.classname);
+  writeNotNull('c__', instance.c__);
   writeNotNull('id', instance.id);
   writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('device_info', instance.deviceInfo);
@@ -114,11 +114,11 @@ Map<String, dynamic> _$GyroscopeDatumToJson(GyroscopeDatum instance) {
 
 LightDatum _$LightDatumFromJson(Map<String, dynamic> json) {
   return LightDatum(
-      meanLux: json['avg_lux'] as num,
+      meanLux: json['mean_lux'] as num,
       stdLux: json['std_lux'] as num,
       minLux: json['min_lux'] as num,
       maxLux: json['max_lux'] as num)
-    ..classname = json['classname'] as String
+    ..c__ = json['c__'] as String
     ..id = json['id'] as String
     ..timestamp = json['timestamp'] == null
         ? null
@@ -137,11 +137,11 @@ Map<String, dynamic> _$LightDatumToJson(LightDatum instance) {
     }
   }
 
-  writeNotNull('classname', instance.classname);
+  writeNotNull('c__', instance.c__);
   writeNotNull('id', instance.id);
   writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('device_info', instance.deviceInfo);
-  writeNotNull('avg_lux', instance.meanLux);
+  writeNotNull('mean_lux', instance.meanLux);
   writeNotNull('std_lux', instance.stdLux);
   writeNotNull('min_lux', instance.minLux);
   writeNotNull('max_lux', instance.maxLux);
@@ -153,7 +153,7 @@ SensorMeasure _$SensorMeasureFromJson(Map<String, dynamic> json) {
       name: json['name'],
       frequency: json['frequency'],
       duration: json['duration'])
-    ..classname = json['classname'] as String
+    ..c__ = json['c__'] as String
     ..enabled = json['enabled'] as bool
     ..configuration = (json['configuration'] as Map<String, dynamic>)
         ?.map((k, e) => MapEntry(k, e as String));
@@ -168,7 +168,7 @@ Map<String, dynamic> _$SensorMeasureToJson(SensorMeasure instance) {
     }
   }
 
-  writeNotNull('classname', instance.classname);
+  writeNotNull('c__', instance.c__);
   writeNotNull('measure_type', instance.measureType);
   writeNotNull('name', instance.name);
   writeNotNull('enabled', instance.enabled);