+++ carp_mobile_sensing/lib/runtime/runtime.dart
@@ -11,13 +11,10 @@
 library runtime;
 
 import 'dart:async';
-import 'dart:io';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:carp_mobile_sensing/probes/sensors/sensors.dart';
-import 'package:device_info/device_info.dart';
-import 'package:flutter/services.dart';
+import 'package:carp_core/carp_core.dart';
 
 part 'probes.dart';
 part 'probe_registry.dart';
 part 'executors.dart';
-part 'device_info.dart';