+++ carp_firebase_backend/CHANGELOG.md
@@ -1,7 +1,11 @@
-## 0.1.0 - Initial version, Oct. 2018.
+## 0.1.1 
+* Bug fixes
+* using `carp_core` as domain model
+
+## 0.1.0 
 * Version tested and ready for use
 * Detailed documentation (see readme.md file)
 * Improved cooperation with FileDataManager
 
-## 0.0.1 - Initial version, Sep. 2018.
+## 0.0.1
 * Initial version of dedicated package bringing together the old implementation.